+++ src/lib/Kakaomap/MapInfoWindow.tsx
@@ -34,6 +34,11 @@ interface MapInfoWindowProps {
    * 로드뷰 상에서 인포윈도우의 가시반경(m 단위), 두 지점 사이의 거리가 지정한 값보다 멀어지면 인포윈도우는 보이지 않게 된다
    */
   range?: number;
+
+  /**
+   * 인포윈도우 객체 생성후 해당 객체를 반환하는 함수
+   */
+  onInfoWindowCreated?: (infoWindow: kakao.maps.InfoWindow) => void;
 }
 
 const MapInfoWindow: React.FC<MapInfoWindowProps> = ({
@@ -44,6 +49,7 @@ const MapInfoWindow: React.FC<MapInfoWindowProps> = ({
   range,
   removable,
   zIndex,
+  onInfoWindowCreated,
 }) => {
   const map = useContext(KakaoMapContext);
 
@@ -60,6 +66,7 @@ const MapInfoWindow: React.FC<MapInfoWindowProps> = ({
       zIndex={zIndex}
       map={map}
       position={infoPosition}
+      onInfoWindowCreated={onInfoWindowCreated}
     >
       {children}
     </InfoWindow>