+++ src/lib/Kakaomap/hooks/index.tsx
@@ -1,6 +1,5 @@
 import useKakaoMapsSDK from "./useKakaoMapsSDK";
-import useMapSetCenter from "./useMapSetCenter";
-import useMapPanTo from "./useMapPanTo";
+import useMapPosition from "./useMapPosition";
 import useRelayout from "./useRelayout";
 
-export { useKakaoMapsSDK, useMapSetCenter, useMapPanTo, useRelayout };
+export { useKakaoMapsSDK, useMapPosition, useRelayout };